windows-cross: cross-build the slic3r-ffi crate via cargo xwin
`cargo xwin build --target x86_64-pc-windows-msvc` now produces a windows
.exe (introspect.exe) that imports slic3r_ffi.dll — the whole C++<->Rust
seam crosses end to end.

build.rs: a Windows branch driving the clang-cl cross toolchain + the
cross-deps prefix (WINCROSS_PREFIX), applying the OrcaSlicer source patches
in place (idempotent), single-config Ninja, dropping the GNU -Wl,-rpath
(lld-link can't parse it), and copying slic3r_ffi.dll beside the
test/example/bin output (Windows resolves DLLs from the exe dir). The Linux
path is unchanged.

lib.rs: cast `as u32` at the two FFI enum boundaries (OptType::Unknown,
OptScope consts). SLIC3R_SCOPE_*/SLIC3R_OPT_* are u32 under GCC but i32
under MSVC — C leaves an all-non-negative enum's underlying type to the
compiler; the values are small positive bitflags so the cast is lossless.
The header is otherwise ABI-clean (only stddef/stdint, no long/wchar).

Linux build + the 26 FFI tests stay green.

+++ crates/slic3r-ffi/src/lib.rs
@@ -166,7 +166,10 @@ impl OptType {
             sys::SLIC3R_OPT_BOOLS => Self::Bools,
             sys::SLIC3R_OPT_ENUM => Self::Enum,
             sys::SLIC3R_OPT_ENUMS => Self::Enums,
-            other => Self::Unknown(other),
+            // The raw enum value is u32 under GCC (Linux) but i32 under MSVC
+            // (windows-msvc), since C leaves an all-non-negative enum's underlying
+            // type to the compiler. Normalize to the wrapper's u32.
+            other => Self::Unknown(other as u32),
         }
     }
 
@@ -214,13 +217,17 @@ impl OptMode {
 pub struct OptScope(pub u32);
 
 impl OptScope {
-    pub const PRINT: Self = Self(sys::SLIC3R_SCOPE_PRINT);
-    pub const OBJECT: Self = Self(sys::SLIC3R_SCOPE_OBJECT);
-    pub const REGION: Self = Self(sys::SLIC3R_SCOPE_REGION);
-    pub const SLA_PRINT: Self = Self(sys::SLIC3R_SCOPE_SLA_PRINT);
-    pub const SLA_OBJECT: Self = Self(sys::SLIC3R_SCOPE_SLA_OBJECT);
-    pub const SLA_MATERIAL: Self = Self(sys::SLIC3R_SCOPE_SLA_MATERIAL);
-    pub const SLA_PRINTER: Self = Self(sys::SLIC3R_SCOPE_SLA_PRINTER);
+    // `as u32`: the SLIC3R_SCOPE_* enum constants are u32 under GCC (Linux) but
+    // i32 under MSVC (windows-msvc) — C leaves an all-non-negative enum's
+    // underlying type to the compiler. The values are small positive bitflags,
+    // so the cast is lossless on both.
+    pub const PRINT: Self = Self(sys::SLIC3R_SCOPE_PRINT as u32);
+    pub const OBJECT: Self = Self(sys::SLIC3R_SCOPE_OBJECT as u32);
+    pub const REGION: Self = Self(sys::SLIC3R_SCOPE_REGION as u32);
+    pub const SLA_PRINT: Self = Self(sys::SLIC3R_SCOPE_SLA_PRINT as u32);
+    pub const SLA_OBJECT: Self = Self(sys::SLIC3R_SCOPE_SLA_OBJECT as u32);
+    pub const SLA_MATERIAL: Self = Self(sys::SLIC3R_SCOPE_SLA_MATERIAL as u32);
+    pub const SLA_PRINTER: Self = Self(sys::SLIC3R_SCOPE_SLA_PRINTER as u32);
 
     /// True if `other`'s bits are all set on `self`.
     pub fn contains(self, other: Self) -> bool {
